requirements.txt
fastapi
strawberry-graphql[fastapi]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-dotenv
uvicorn
requests
alembic